Increase the network capacity over 256 peers
While improbable, if a chain of more than 256 badges is created the peer
id whould overflow a uint8_t, increase it to uint16_t.

+++ components/badge-network/network_handler.cpp
@@ -228,11 +228,11 @@ struct wire_msg_header {
 } __attribute__((packed));
 
 struct wire_msg_announce {
-    std::uint8_t peer_id;
+    std::uint16_t peer_id;
 } __attribute__((packed));
 
 struct wire_msg_announce_reply {
-    std::uint8_t peer_count;
+    std::uint16_t peer_count;
 } __attribute__((packed));
 
 std::uint8_t wire_msg_payload_size(std::uint8_t type)
